feat: 개별 종목 뉴스 목록 조회
S12P21A304-97

+++ backend/src/main/java/newstock/domain/news/entity/News.java
@@ -14,7 +14,7 @@ public class News {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int newsId;
+    private long newsId;
 
     private int stockCode;
